feat/#75: 멤버 도메인으로 이동, 이미지는 /api/v1/images/upload/type/{imageType}로 관리
