--- src/main/java/com/divary/domain/mypage/dto/response/MyPageImageResponseDTO.java
@@ -1,12 +1,0 @@
-package com.divary.domain.mypage.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@AllArgsConstructor
-@NoArgsConstructor
-public class MyPageImageResponseDTO {
-    private String url;
-}
